Revert "Merge branch 'main' into pr/1380"
This reverts commit 29ef2838553003ad0534edc821038ec26c3f2e28, reversing
changes made to de564fe3879121c708fbd5a52058443416b8e7ec.

+++ src/framework/otml/otmlnode.h
@@ -33,7 +33,7 @@ public:
     static OTMLNodePtr create(std::string_view tag, std::string_view value);
 
     std::string tag() { return m_tag; }
-    size_t size() const { return m_children.size(); }
+    int size() const { return m_children.size(); }
     std::string source() { return m_source; }
     std::string rawValue() { return m_value; }
